models directory w/ new separated files
