+++ form_builder/models/__init__.py
@@ -1,0 +1,2 @@
+from .fastq import Fastq, SingleEndFastq, PairedEndFastq
+from .sample import Sample
